Refactor shots to their own source file.
