+++ Makefile
@@ -1,5 +1,5 @@
 PRJNAME := dragon_blaster
-OBJS := data.rel actor.rel shot.rel map.rel score.rel dragon_blaster.rel
+OBJS := data.rel actor.rel shot.rel shots.rel map.rel score.rel dragon_blaster.rel
 
 all: $(PRJNAME).sms
